+++ CHANGELOG.md
@@ -5,9 +5,13 @@ Unreleased
 ----------
 ### Changed
 * Add wdio tests for additional themes, update travis ci config.
-
-### Changed
 * Add additional keyboard navigation options to CollapsingNavigationMenu
+
+### Removed
+* Removed toggle backgrounds tab and functionality from example loader
+
+### Added
+* Added CSS display tab and functionality to example loader
 
 6.25.0 - (July 7, 2020)
 ----------